**Summary**

- Refactor Triton kernels to use a different approach for calculating the global amax
- Have a single shared global amax tensor of size 1
- All blocks compute local amax then perform a thread-safe max operation `tl.atomic_max` with the shared global max tensor to find the true global max once the kernel is complete
- The advantage of this strategy is we don't have to allocate a shared buffer of size `num_elements // BLOCK_SIZE` outside of the kernels, which prevents us from easily autotuning the block size

**Test plan**
- `pytest test/test.py` - Passing
- `python3 benchmark/benchmark.py` yields large performance improvement as seen below. It's now markedly better than production path eager execution but not quite as good as compiled yet.

Old performance with un-tuned block size of 8:
```
  input_size  high_precision_dtype      eager_time    compiled_time    float8nocompile
-------------  ----------------------  ------------  ---------------  -----------------
 65500         torch.float32                599.299          298.101    94446
 65500         torch.bfloat16               649.674          394.535    94386.3
     1.05e+06  torch.float32                640.5            332.171   104449
     1.05e+06  torch.bfloat16               685.822          421.365   104372
     1.68e+07  torch.float32               1963.09          1214.32    280825
     1.68e+07  torch.bfloat16              1828.16          1051.67    261710
     2.68e+08  torch.float32              24129.8          16287.2          3.39791e+06
     2.68e+08  torch.bfloat16             21603.2          12389.9          3.39515e+06
```

New performance with auto-tuned block size:
```
  input_size  high_precision_dtype      eager_time    compiled_time    float8nocompile
------------  ----------------------  ------------  ---------------  -----------------
65500         torch.float32                599.055          298.543            372.018
65500         torch.bfloat16               649.523          394.457            413.137
    1.05e+06  torch.float32                640.497          332.419            413.503
    1.05e+06  torch.bfloat16               685.584          421.296            453.472
    1.68e+07  torch.float32               1963.72          1215.19            1415.5
    1.68e+07  torch.bfloat16              1829.28          1051.86            1298.55
    2.68e+08  torch.float32              24126.2          16294.3            19124.8
    2.68e+08  torch.bfloat16             21592.5          12390.6            16485.7
```