@danielvegamyhre 2024-12-20
> Looks good. I'm not sure I would completely throw away the old though since I think tuning the split reduction could get us a lot of performance and is deterministic which is nice

@drisspg I updated the PR to keep both implementation and make the kernel algorithm used configurable, and updated unit tests to exercise both paths.

One thing I noticed though is that the atomic max strategy was failing non-deterministically, I had to add a call ` torch.cuda.synchronize()` after the global amax kernel to fix this.