@vkuzo 2024-12-23
> Performance of atomic max based kernel with auto-tuned block size:

A few comments about measurement:
1. feel free to only report bfloat16 to keep things simple
2. for `input_size`, it would be good to report the 3d tensor shape instead of the 1d number of elements, to make it easy for the reader to map back to real tensor shapes in common models