@danielvegamyhre 2024-12-23
> > Performance of atomic max based kernel with auto-tuned block size:
> 
> A few comments about measurement:
> 
> 1. feel free to only report bfloat16 to keep things simple
> 2. for `input_size`, it would be good to report the 3d tensor shape instead of the 1d number of elements, to make it easy for the reader to map back to real tensor shapes in common models

Makes sense, done! New example output:

```
input_shape    high_precision_dtype      eager_time    compiled_time    float8nocompile
-------------  ----------------------  ------------  ---------------  -----------------
(16, 4096)     torch.bfloat16               649.843          395.745            412.926
(256, 4096)    torch.bfloat16               685.372          421.237            449.762
(4096, 4096)   torch.bfloat16              1828.54          1053.8             1244.41
(65536, 4096)  torch.bfloat16             21576            12438.9            15619.6
```